Update readme.md
the examples are wrong, they both return "1, 2, 3" when instead the first one should return "3, 3, 3" and the second one "1, 1, 1". The error was the use of await in the log function which would prevent pDebounce from doing its work. The fix was to replace with the then() statement to not block pDebounce